* 'master' of git://scm.dev.nokia.troll.no/qt/qt-doc-team: (110 commits)
  Doc: adding link to the Qt Quick Components for Symbian page.
  Update Japanese translations.
  Updated Russian translation
  Update Korean translations to 4.8
  qmake: fix writing of incremental linker option for vcxproj files
  support the obj-c #import statement in the dependency generator
  runonphone timestamps nanosecond formating set to 9 chars, 0 padded
  Put Qt translation files to rom on Symbian.
  Add -nn for .cpp MOC preprocessing for INTEGRITY
  Remove commented out code.
  Update Polish translations for 4.8
  Fix QDeclarativeEngine::setOfflineStoragePath() for Symbian
  Fixes: libpng symbols exported from QtGui.dll on Windows
  Converting from double to qreal in gui
  Doc updates to installation, platform notes and symbian introduction.
  Fix XML schema validation failure.
  omit unassigned (and too recent codepoints) from the text Normalization process
  purge "Translation Rules for Plurals"
  add -list-languages option to lupdate
  Fix construction races in QtNetwork
  ...

+++ src/gui/graphicsview/qgraphicswidget_p.cpp
@@ -482,8 +482,8 @@ static QSizeF closestAcceptableSize(const QSizeF &proposed,
             minw = maxw;
             minh = maxh;
         }
-        middlew = minw + (maxw - minw)/2.0;
-        middleh = minh + (maxh - minh)/2.0;
+        middlew = minw + (maxw - minw)/qreal(2.0);
+        middleh = minh + (maxh - minh)/qreal(2.0);
 
         min_hfw = minimumHeightForWidth(middlew, minh, maxh, widget);
